Start with index.html
Q: Hi, please show me how to create a Web application that will show today's weather forecast, relative to yesterday.

A:
Sure, here is an outline of the steps you would need to follow to create a simple web application that displays today's weather forecast relative to yesterday's:

First, you will need to decide on the technology you want to use to build your web application. Some popular options for creating web applications include using a framework like Ruby on Rails, Django, or Flask, or using a content management system like WordPress or Drupal.

Once you have chosen the technology you want to use, you will need to set up a web server to host your application. You can either set up your own web server, or use a cloud hosting service like Amazon Web Services, Microsoft Azure, or Google Cloud Platform.

Next, you will need to create the user interface for your application. This will involve designing the layout and appearance of your application, as well as creating any necessary HTML, CSS, and JavaScript code.

To display the weather forecast in your application, you will need to use a weather API (Application Programming Interface) to retrieve the data. There are many weather APIs available, some of which are free to use and others which require a subscription. Once you have signed up for a weather API, you can use its documentation to learn how to retrieve the data you need and display it in your application.

Finally, you will need to write the code for your application to retrieve the current weather forecast from the API and display it in the user interface. This will involve writing code to make requests to the API and process the response, as well as code to display the data in the appropriate format in your application.

Overall, creating a web application that shows the weather forecast relative to yesterday's can be a challenging but rewarding project. By following the steps outlined above, you can create a simple but functional weather forecast web application.

Q: I want to do that without a server, just a HTML + JavaScript page

A:
Yes, it is possible to create a simple weather forecast web application using only HTML and JavaScript. Here are the steps you would need to follow:

First, you will need to create an HTML page for your application. This will involve writing the necessary HTML code to create the layout and structure of your page.

Next, you will need to use a weather API to retrieve the current weather forecast data. As with the previous example, there are many weather APIs available, and you will need to choose one and sign up for an API key. Once you have your API key, you can use the API's documentation to learn how to make requests to the API and retrieve the data you need.

To make requests to the API and retrieve the weather data, you will need to use JavaScript. You can use the fetch function to make HTTP requests to the API and retrieve the data, and then use the JSON.parse function to parse the response and access the data you need.

Once you have retrieved the weather data from the API, you will need to display it in your HTML page. You can use JavaScript to access and manipulate the page's DOM (Document Object Model) to add the weather data to the page.

Finally, you will need to write the code for your application to compare today's weather forecast to yesterday's. This will involve retrieving the weather data for both days, comparing the data, and displaying the results in your HTML page.

Overall, creating a simple weather forecast web application using only HTML and JavaScript is a feasible project. By following the steps outlined above, you can create a functional weather forecast application that runs in the user's web browser without the need for a server.

Q: Let's start with 1. Can you show me the code?

A:
Sure, here is an example of an HTML page that could be used as the basis for a weather forecast application:

<!DOCTYPE html>
<html>
  <head>
    <title>Weather Forecast</title>
  </head>
  <body>
    <h1>Weather Forecast</h1>
    <div id="forecast"></div>
  </body>
</html>

This HTML page includes a <div> element with an id attribute of "forecast". This div element will be used to display the weather data retrieved from the API.

Note that this is just an example, and you can customize the HTML code to suit your needs and preferences. For example, you could add additional elements and formatting to the page to create the desired layout and appearance for your application.