Q: Now create a README file for this project
A:

Sure, here is an example of a README file that could be used for a weather forecast web application:

Weather Forecast
This is a simple web application that shows the current weather forecast for a given city. The application retrieves the forecast data from a weather API and displays it in the user's web browser.

Getting Started
These instructions will get you a copy of the project up and running on your local machine for development and testing purposes.

Prerequisites
To run this application, you will need:

A web browser
An internet connection
An API key for a weather API (e.g. openweathermap.org)
Installing
To install the application, follow these steps:

Clone or download the repository to your local machine.
Open the index.html file in your web browser.
Replace the API_KEY placeholder in the main.js file with your own API key.
Usage
To use the application, simply open the index.html file in your web browser. The current weather forecast for the city specified in the code will be displayed on the page.

Built With
HTML
JavaScript
openweathermap.org - The weather API used to retrieve the forecast data
Authors
Your Name - Initial work
License
This project is licensed under the MIT License - see the LICENSE file for details.

Acknowledgments
openweathermap.org for providing the weather data used in this application.
Keywords
weather
forecast
web application
HTML
JavaScript
API

Q: The description is not correct. This is a simple web app that shows the current weather compared to yestertday

A:

Weather Comparison
This is a simple web application that shows the current weather compared to yesterday's forecast for a given city. The application retrieves the forecast data from a weather API and displays it in the user's web browser.
.
.
.

Q: Now just add a sentence about this app was built step by step with you

A:
Sure, here is an updated version of the README file with the additional sentence:

Weather Comparison
This is a simple web application that shows the current weather compared to yesterday's forecast for a given city. The application was built step by step with the help of a large language model trained by OpenAI. The application retrieves the forecast data from a weather API and displays it in the user's web browser.

Getting Started