Request yesterday weather in cloud function
Q: Use "openweathermap" in Firebase function to get the weather of yesterday.

A:
To get the weather of yesterday using the openweathermap API in a Firebase function, you can follow these steps:

Install the axios package in your Firebase project by running the following command in your project directory:
Copy code
npm install axios
Import the axios package in your Firebase function:
javascript
Copy code
const axios = require('axios');
Define the getWeather function to retrieve the weather data from the openweathermap API. The function should take the latitude and longitude of the location for which you want to get the weather data, as well as a date parameter representing the date for which you want to get the weather data. The function should use the axios package to make a GET request to the openweathermap API, passing in the latitude, longitude, date, and API key as query parameters. The response from the API will include an array of weather data for the specified location and date. You can then filter this array to find the weather data for the previous day:

async function getWeather(latitude, longitude, date) {
  const apiKey = '<your-openweathermap-api-key>';
  const apiUrl = `https://api.openweathermap.org/data/2.5/onecall/timemachine?lat=${latitude}&lon=${longitude}&dt=${date}&appid=${apiKey}`;
  const response = await axios.get(apiUrl);
  const weatherData = response.data.hourly;
  const yesterdayWeather = weatherData.filter((weather) => {
    const weatherDate = new Date(weather.dt * 1000);
    const yesterday = new Date(date * 1000);
    yesterday.setDate(yesterday.getDate() - 1);
    return weatherDate.getDate() === yesterday.getDate();
  })[0];
  return yesterdayWeather;
}
Call the getWeather function in your Firebase function, passing in the latitude, longitude, and date for which you want to get the weather data:

exports.getYesterdayWeather = functions.https.onRequest(async (req, res) => {
  const latitude = <latitude>;
  const longitude = <longitude>;
  const date = <yesterday's date in Unix timestamp format>;
  const yesterdayWeather = await getWeather(latitude, longitude, date);
  res.send(yesterdayWeather);
});
Note that you'll need to replace <your-openweathermap-api-key>, <latitude>, <longitude>, and <yesterday's date in Unix timestamp format> with your own values. You can get an API key by signing up for an openweathermap account. The latitude and longitude can be obtained from a location service or hardcoded if you know the location for which you want to get the weather data. The date parameter should be set to the Unix timestamp representing yesterday's date.

Q: Use TypeScript

A:
Sure! Here's how you can write the same Firebase function in TypeScript:

import * as functions from 'firebase-functions';
import * as admin from 'firebase-admin';
import axios from 'axios';

admin.initializeApp();

async function getYesterdayWeather(latitude: number, longitude: number, date: number) {
  const apiKey = '<your-openweathermap-api-key>';
  const apiUrl = `https://api.openweathermap.org/data/2.5/onecall/timemachine?lat=${latitude}&lon=${longitude}&dt=${date}&appid=${apiKey}`;
  const response = await axios.get(apiUrl);
  const weatherData = response.data.hourly;
  const yesterdayWeather = weatherData.filter((weather: { dt: number }) => {
    const weatherDate = new Date(weather.dt * 1000);
    const yesterday = new Date(date * 1000);
    yesterday.setDate(yesterday.getDate() - 1);
    return weatherDate.getDate() === yesterday.getDate();
  })[0];
  return yesterdayWeather;
}

export const getYesterdayWeather = functions.https.onRequest(async (req, res) => {
  const latitude = <latitude>;
  const longitude = <longitude>;
  const date = <yesterday's date in Unix timestamp format>;
  const yesterdayWeather = await getYesterdayWeather(latitude, longitude, date);
  res.send(yesterdayWeather);
});
Make sure you have the necessary packages installed by running:

npm install firebase-functions firebase-admin axios
And also make sure to replace <your-openweathermap-api-key>, <latitude>, <longitude>, and <yesterday's date in Unix timestamp format> with your own values.

Q: Why not calculating the date of yesterday from the current day?

A:
Calculating the date of yesterday from the current day is also a valid approach, and it can be simpler and more concise than the approach I described earlier.

Here's how you can modify the getYesterdayWeather function to calculate the date of yesterday from the current day:

async function getYesterdayWeather(latitude: number, longitude: number) {
  const apiKey = '<your-openweathermap-api-key>';
  const apiUrl = `https://api.openweathermap.org/data/2.5/onecall?lat=${latitude}&lon=${longitude}&exclude=current,minutely,daily&appid=${apiKey}`;
  const response = await axios.get(apiUrl);
  const yesterdayIndex = response.data.hourly.length - 25; // 25 hours ago
  const yesterdayWeather = response.data.hourly[yesterdayIndex];
  return yesterdayWeather;
}
In this approach, we're getting the current weather data for the specified location using the onecall endpoint of the openweathermap API. We're excluding the current, minutely, and daily weather data, and only getting the hourly weather data.

The response from the API will include an array of hourly weather data for the next 48 hours, starting from the current hour. We can calculate the index of the hourly weather data for yesterday by subtracting 25 from the length of the array. This is because 25 hours ago would be the last hour of yesterday.

Note that this approach assumes that you're calling the Firebase function at a time when the current hour is at least 1 hour into the current day. If you call the function before the first hour of the current day has passed, the weather data for yesterday will not be available yet.

Also note that this approach doesn't require the date parameter, since we're calculating the date of yesterday based on the current date and time.

Q: Get the latitude & longitude from the request fields

A:
Sure, here's how you can modify the Firebase function to get the latitude and longitude values from the request fields:

export const getYesterdayWeather = functions.https.onRequest(async (req, res) => {
  const { latitude, longitude } = req.query;
  if (!latitude || !longitude) {
    res.status(400).send('Missing latitude or longitude');
    return;
  }
  const yesterdayWeather = await getYesterdayWeather(parseFloat(latitude), parseFloat(longitude));
  res.send(yesterdayWeather);
});
In this modified version of the function, we're getting the latitude and longitude values from the query object of the req parameter. We're checking if either of these values is missing, and returning a 400 Bad Request response if that's the case.

We're then converting the latitude and longitude values to numbers using the parseFloat function, and passing them to the getYesterdayWeather function to get the weather data for yesterday. Finally, we're sending the yesterdayWeather object as the response.